feat(api): add hotkey callback registration system
