+++ src/ClassicUO.Client/Input/Keyboard.cs
@@ -1,6 +1,7 @@
 ﻿// SPDX-License-Identifier: BSD-2-Clause
 
 using System;
+using System.Collections.Generic;
 using SDL3;
 
 namespace ClassicUO.Input
@@ -12,21 +13,52 @@ namespace ClassicUO.Input
         public static bool Alt { get; private set; }
         public static bool Shift { get; private set; }
         public static bool Ctrl { get; private set; }
-        public static event Action<SDL.SDL_KeyboardEvent> KeyDownEvent;
-        public static event Action<SDL.SDL_KeyboardEvent> KeyUpEvent;
+        public static event Action<string> KeyDownEvent;
+        public static event Action<string> KeyUpEvent;
+
+        public static bool IgnoreBareModifierKey(SDL.SDL_KeyboardEvent e)
+        {
+            var keycode = (SDL.SDL_Keycode)e.key;
+            return keycode is SDL.SDL_Keycode.SDLK_LSHIFT
+                        or SDL.SDL_Keycode.SDLK_RSHIFT
+                        or SDL.SDL_Keycode.SDLK_LCTRL
+                        or SDL.SDL_Keycode.SDLK_RCTRL
+                        or SDL.SDL_Keycode.SDLK_LALT
+                        or SDL.SDL_Keycode.SDLK_RALT;
+        }
+
+        public static string BuildHotKeyString(SDL.SDL_KeyboardEvent e)
+        {
+            List<string> parts = new();
+            if (Ctrl) parts.Add("CTRL");
+            if (Shift) parts.Add("SHIFT");
+            if (Alt) parts.Add("ALT");
+
+            string keyName = ((SDL.SDL_Keycode)e.key).ToString().ToUpperInvariant();
+            parts.Add(keyName);
+
+            return string.Join("+", parts);
+        }
+
 
         public static void OnKeyUp(SDL.SDL_KeyboardEvent e)
         {
-            UpdateModifiers(e.mod);
-            KeyUpEvent?.Invoke(e);
+            OnKeyEvent(e, KeyUpEvent);
         }
 
         public static void OnKeyDown(SDL.SDL_KeyboardEvent e)
         {
-            UpdateModifiers(e.mod);
-            KeyDownEvent?.Invoke(e);
+            OnKeyEvent(e, KeyDownEvent);
         }
 
+        private static void OnKeyEvent(SDL.SDL_KeyboardEvent e, Action<string> keyboardEvent)
+        {
+            UpdateModifiers(e.mod);
+            if (IgnoreBareModifierKey(e))
+                return;
+            string hotkey = BuildHotKeyString(e);
+            keyboardEvent?.Invoke(hotkey);
+        }
 
         private static void UpdateModifiers(SDL.SDL_Keymod e)
         {
